ludika-backend/pyproject.toml
[tool.poetry]
name = "ludika-backend"
version = "0.1.0"
description = ""
authors = ["TheManchineel <37479927+TheManchineel@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
pydantic = "^2.11.3"
sqlmodel = "^0.0.24"
psycopg = "^3.2.6"
uvicorn = "^0.34.2"
requests = "^2.32.3"
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
authlib = "^1.5.2"
pyjwt = "^2.10.1"
itsdangerous = "^2.2.0"
google-auth = "^2.40.1"
httpx = "^0.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
